tidy up mass valume code
Plus copy original voxel obj
