--- tutorials/get_mass/get_mass.go
@@ -1,62 +1,0 @@
-package main
-
-import (
-	"bufio"
-	"encoding/base64"
-	"encoding/json"
-	"fmt"
-	"github.com/kittycad/kittycad.go"
-	"io/ioutil"
-	"os"
-	"strings"
-)
-
-func main() {
-	// Create a new client with your token parsed from the environment variable:
-	// KITTYCAD_API_TOKEN.
-	client, err := kittycad.NewClientFromEnv("your apps user agent")
-	if err != nil {
-		panic(err)
-	}
-
-	input, _ := os.Open("./ORIGINALVOXEL-3.obj")
-
-	content, _ := ioutil.ReadAll(bufio.NewReader(input))
-
-	// Encode as base64.
-	myReader := strings.NewReader(base64.StdEncoding.EncodeToString(content))
-
-	densitySteelGramsPerCubicMillimeter := 0.00784
-
-	fc, err := client.File.CreateMass(
-		densitySteelGramsPerCubicMillimeter,
-		"obj",
-		myReader,
-	)
-	if err != nil {
-		panic(err)
-	}
-
-	fmt.Println("File mass id: ", fc.ID)
-	fmt.Println("File mass (g): ", fc.Mass)
-
-	json_data, err := json.Marshal(fc)
-	if err != nil {
-		panic(err)
-	}
-	output_file_path := "./output.json"
-
-	output, err := os.Create(output_file_path)
-	if err != nil {
-		panic(err)
-	}
-
-	defer output.Close()
-
-	if _, err := output.Write(json_data); err != nil {
-		panic(err)
-	}
-	if err := output.Sync(); err != nil {
-		panic(err)
-	}
-}
