--- tutorials/get_mass/output.json
@@ -1,5 +1,0 @@
-{
-    "id": "dd291644-db73-4395-a72d-07eae6f3d26c",
-    "status": "Completed",
-    "mass": 0.79215306
-}
