+++ tutorials/get_mass_volume/get_mass_volume.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python3
 from kittycad.client import ClientFromEnv
-from kittycad.models import file_source_format
-from kittycad.api.file import create_file_volume
-from kittycad.models import FileVolume
+from kittycad.models import file_source_format, FileMass, FileVolume
+from kittycad.api.file import create_file_mass, create_file_volume
 import json
+import os
 
 # Create a new client with your token parsed from the environment variable:
 #   KITTYCAD_API_TOKEN.
@@ -14,19 +14,27 @@ file = open("./ORIGINALVOXEL-3.obj", "rb")
 content = file.read()
 file.close()
 
+steelDensityPerCubicMillimeter = 0.00785
+fm: FileMass = create_file_mass.sync(
+    client=client,
+    material_density=steelDensityPerCubicMillimeter,
+    src_format=file_source_format.FileSourceFormat.OBJ,
+    body=content)
 fv: FileVolume = create_file_volume.sync(
     client=client,
     src_format=file_source_format.FileSourceFormat.OBJ,
     body=content)
 
+print(f"File mass (grams): {fm.mass}")
 print(f"File volume (mm³): {fv.volume}")
 
 with open('output.json', 'w', encoding='utf-8') as f:
     json.dump({
-        "id": fv.id,
-        "status": fv.status,
+        "ids": [fm.id, fv.id],
         "volume": fv.volume,
+        "mass": fm.mass,
     },
               f,
               ensure_ascii=False,
               indent=4)
+os.system("cp ./ORIGINALVOXEL-3.obj ./output.obj")
