--- tutorials/get_volume/get_volume.json
@@ -1,15 +1,0 @@
-{
-  "name": "title",
-  "outputPath": "tutorials/get_volume/output.json",
-  "outputFormat": "json",
-  "files": [
-    {
-      "extension": "go",
-      "range": [22, 17]
-    },
-    {
-      "extension": "py",
-      "range": [16, 10]
-    }
-  ]
-}
