--- tutorials/get_volume/output.json
@@ -1,5 +1,0 @@
-{
-    "id": "e43369c5-ae8f-4ae2-bab8-d9ae323663de",
-    "status": "Completed",
-    "volume": 100.911224
-}
